requirements.txt
# All dependencies (for development)

# Backend
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
httpx>=0.26.0,<1.0.0
aiosqlite>=0.19.0,<1.0.0
minecraft-launcher-lib>=5.0.0,<6.0.0
cryptography>=42.0.0,<43.0.0

# Frontend
PyQt6>=6.6.0,<7.0.0
PyQt6-Qt6>=6.6.0,<7.0.0

# Common
aiohttp>=3.9.0,<4.0.0
loguru>=0.7.0,<1.0.0
requests>=2.31.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
websockets>=12.0,<13.0
websocket-client>=1.6.0

# Microsoft Auth
msal>=1.28.0,<2.0.0

# Server Ping
mcstatus>=11.1.0,<12.0.0

# Local LLM
llama-cpp-python>=0.2.0

# Testing (optional)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
